pix_control_api/requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
pyjwt==2.8.0
python-dotenv==1.0.0
python-multipart==0.0.6
colorlog==6.9.0
httpx==0.28.1
requests==2.32.3